# kardo-swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "KARDO API"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/comments/videos/{videoId}": {
      "get": {
        "tags": [
          "Комментарии."
        ],
        "summary": "Получение списка комментариев к видео-клипу.",
        "operationId": "getCommentsByVideoId",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "id видео-клипа",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Количество элементов, которые нужно пропустить для формирования текущего набора",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Количество элементов в наборе",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-клип не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Список комментариев пользователя получен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Комментарии."
        ],
        "summary": "Добавление нового комментария к видео-клипу.",
        "operationId": "createComment",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "id видео-клипа",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-клип не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Комментарий добавлен.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/contests": {
      "post": {
        "tags": [
          "Видео-соревнование: Admin."
        ],
        "summary": "Добавление видео-соревнования.",
        "operationId": "createVideoContest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewVideoContestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Пользователь не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Видео-соревнование добавлено.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoContestDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/videos/{videoId}": {
      "get": {
        "tags": [
          "Видео-клип."
        ],
        "summary": "Получение видео-клипа.",
        "operationId": "getVideoClipById",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "id видео-клипа",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-клип не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Видео-клип найден.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoClipDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Видео-клип."
        ],
        "summary": "Удаление видео-клипа.",
        "operationId": "deleteVideoClipById",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "id видео-клипа",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-клип не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Видео-клип удален."
          }
        }
      },
      "patch": {
        "tags": [
          "Видео-клип."
        ],
        "summary": "Обновление видео-клипа.",
        "operationId": "updateVideoClipById",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "id видео-клипа",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVideoClipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-клип не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Видео-клип обновлён.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoClipDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/selections/video/{selectionId}": {
      "delete": {
        "tags": [
          "Видео-отбор: Admin."
        ],
        "summary": "Удаление видео-отбора.",
        "operationId": "deleteVideoSelection",
        "parameters": [
          {
            "name": "selectionId",
            "in": "path",
            "description": "id видео-отбора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-отбор не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Видео-отбор удалён."
          }
        }
      },
      "patch": {
        "tags": [
          "Видео-отбор: Admin."
        ],
        "summary": "Обновление видео-отбора.",
        "operationId": "updateVideoSelectionById",
        "parameters": [
          {
            "name": "selectionId",
            "in": "path",
            "description": "id видео-отбора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVideoSelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-отбор не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Видео-отбор обновлён.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoSelectionDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/contests/{contestId}": {
      "delete": {
        "tags": [
          "Видео-соревнование: Admin."
        ],
        "summary": "Удаление видео-соревнования.",
        "operationId": "deleteVideoContest",
        "parameters": [
          {
            "name": "contestId",
            "in": "path",
            "description": "id видео-соревнования",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-соревнование не найдено",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Видео-соревнование удалено."
          }
        }
      },
      "patch": {
        "tags": [
          "Видео-соревнование: Admin."
        ],
        "summary": "Обновление видео-соревнования.",
        "operationId": "updateVideoContest",
        "parameters": [
          {
            "name": "contestId",
            "in": "path",
            "description": "id видео-соревнования",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVideoContestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-соревнование не найдено",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Видео-соревнование обновлено.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoContestDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/selections/video/{selectionId}": {
      "get": {
        "tags": [
          "Видео-отбор: Users."
        ],
        "summary": "Получение видео-отбора.",
        "operationId": "getVideoSelectionById",
        "parameters": [
          {
            "name": "selectionId",
            "in": "path",
            "description": "id видео-отбора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-отбор не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Видео-отбор найден.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoSelectionDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/contests/{contestId}": {
      "get": {
        "tags": [
          "Видео-соревнование: Users."
        ],
        "summary": "Получение видео-соревнования.",
        "operationId": "getVideoContestById",
        "parameters": [
          {
            "name": "contestId",
            "in": "path",
            "description": "id видео-соревнования",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Видео-соревнование не найдено",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Видео-соревнование найдено.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoContestDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserShortDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "birthday": {
            "type": "string",
            "format": "date"
          },
          "country": {
            "type": "string"
          },
          "avatarPhoto": {
            "type": "string"
          }
        },
        "description": "Сущность пользователя короткая."
      },
      "VideoClipDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "publicationTime": {
            "type": "string",
            "format": "date-time"
          },
          "hashtags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "creatorDto": {
            "$ref": "#/components/schemas/UserShortDto"
          },
          "likesCount": {
            "type": "integer",
            "format": "int32"
          },
          "videoLink": {
            "type": "string"
          }
        },
        "description": "Сущность видео-клипа."
      },
...
      "GrandCompetitionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "hashtag": {
            "type": "string"
          },
          "competitionStart": {
            "type": "string",
            "format": "date"
          },
          "competitionEnd": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "FINISHED",
              "UPCOMING",
              "IN_PROGRESS"
            ]
          },
          "location": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "BREAKING",
                "PARKOUR",
                "WORKOUT",
                "KICKSCOOTERING",
                "TRICKING",
                "SKATEBOARDING",
                "BMX",
                "HIP_HOP",
                "DJ_ING",
                "GRAFFITI",
                "WAKEBOARDING",
                "ALL_FIELD"
              ]
            }
          },
          "description": {
            "type": "string"
          }
        },
        "description": "Сущность Гранд-соревнования."
      },
...
      "NewCommentRequest": {
        "required": [
          "authorId",
          "text"
        ],
        "type": "object",
        "properties": {
          "authorId": {
            "type": "integer",
            "format": "int64"
          },
          "text": {
            "maxLength": 500,
            "minLength": 2,
            "type": "string"
          }
        },
        "description": "Данные добавляемого комментария"
      },
      "CommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "authorDto": {
            "$ref": "#/components/schemas/UserShortDto"
          },
          "publicationTime": {
            "type": "string",
            "format": "date-time"
          },
          "videoClipDto": {
            "$ref": "#/components/schemas/VideoClipDto"
          },
          "text": {
            "type": "string"
          }
        },
        "description": "Сущность комментария к видео-клипу."
      },
...
      "VideoSelectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "competitionDto": {
            "$ref": "#/components/schemas/GrandCompetitionDto"
          },
          "selectionStart": {
            "type": "string",
            "format": "date"
          },
          "selectionEnd": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "FINISHED",
              "UPCOMING",
              "IN_PROGRESS"
            ]
          },
          "hashtag": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "description": "Сущность видео-отбора."
      },
...
      "NewVideoContestRequest": {
        "required": [
          "contestEnd",
          "contestStart",
          "description",
          "hashtag",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 250,
            "minLength": 2,
            "type": "string"
          },
          "hashtag": {
            "maxLength": 20,
            "minLength": 2,
            "type": "string"
          },
          "contestStart": {
            "type": "string",
            "format": "date"
          },
          "contestEnd": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "maxLength": 3000,
            "minLength": 2,
            "type": "string"
          }
        },
        "description": "Данные добавляемого видео-соревнования"
      },
      "VideoContestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "hashtag": {
            "type": "string"
          },
          "contestStart": {
            "type": "string",
            "format": "date"
          },
          "contestEnd": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "FINISHED",
              "UPCOMING",
              "IN_PROGRESS"
            ]
          },
          "description": {
            "type": "string"
          }
        },
        "description": "Сущность видео-соревнования."
      },
...
      "UpdateVideoClipRequest": {
        "required": [
          "hashtags"
        ],
        "type": "object",
        "properties": {
          "hashtags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "Данные обновляемого видео-клипа"
      },
...
      "UpdateVideoSelectionRequest": {
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 250,
            "minLength": 2,
            "type": "string"
          },
          "selectionStart": {
            "type": "string",
            "format": "date"
          },
          "selectionEnd": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "FINISHED",
              "UPCOMING",
              "IN_PROGRESS"
            ]
          },
          "description": {
            "maxLength": 6500,
            "minLength": 2,
            "type": "string"
          }
        },
        "description": "Данные обновляемого видео-отбора"
      },
...
      "UpdateVideoContestRequest": {
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 250,
            "minLength": 2,
            "type": "string"
          },
          "contestStart": {
            "type": "string",
            "format": "date"
          },
          "contestEnd": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "FINISHED",
              "UPCOMING",
              "IN_PROGRESS"
            ]
          },
          "description": {
            "maxLength": 3000,
            "minLength": 2,
            "type": "string"
          }
        },
        "description": "Данные обновляемого видео-соревнования"
      },
...
    }
  }
}